+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## placeholder
-* Update sdk loggin logics so all level logs are recored into kusto. ([#167](https://github.com/microsoft/durabletask-java/pull/167))
+* Update sdk loggin logics so all level logs are recored into kusto. ([#170](https://github.com/microsoft/durabletask-java/pull/170))
 
 ## v1.4.0